lazy load diagnostic dialog
