+++ ui/site/src/component/log.ts
@@ -1,6 +1,4 @@
 import { objectStorage, ObjectStorage } from 'common/objectStorage';
-import { isTouchDevice } from 'common/device';
-import { domDialog } from 'common/dialog';
 
 export default function makeLog(): LichessLog {
   let store: ObjectStorage<string, number>;
@@ -58,8 +56,6 @@ export default function makeLog(): LichessLog {
     return keys.map((k, i) => `${new Date(k).toISOString()} - ${vals[i]}`).join('\n');
   };
 
-  log.diagnostic = show;
-
   window.addEventListener('error', async e => {
     log(`${e.message} (${e.filename}:${e.lineno}:${e.colno})\n${e.error?.stack ?? ''}`.trim());
   });
@@ -68,33 +64,4 @@ export default function makeLog(): LichessLog {
   });
 
   return log;
-}
-
-async function show() {
-  const logs = await lichess.log.get();
-  const text =
-    `Browser: ${navigator.userAgent}\n` +
-    `Cores: ${navigator.hardwareConcurrency}\n` +
-    `Touch: ${isTouchDevice()} ${navigator.maxTouchPoints}\n` +
-    `Screen: ${window.screen.width}x${window.screen.height}\n` +
-    `Lang: ${navigator.language}` +
-    (logs ? `\n\n${logs}` : '');
-
-  const dlg = await domDialog({
-    class: 'diagnostic',
-    cssPath: 'diagnostic',
-    htmlText:
-      `<h2>Diagnostics</h2><pre tabindex="0" class="err">${lichess.escapeHtml(text)}</pre>` +
-      (logs ? `<button class="clear button">Clear Logs</button>` : ''),
-  });
-  const select = () =>
-    setTimeout(() => {
-      const range = document.createRange();
-      range.selectNodeContents(dlg.view.querySelector('.err')!);
-      window.getSelection()?.removeAllRanges();
-      window.getSelection()?.addRange(range);
-    }, 0);
-  $('.err', dlg.view).on('focus', select);
-  $('.clear', dlg.view).on('click', () => lichess.log.clear().then(lichess.reload));
-  dlg.showModal();
 }
